Moved extensions from ui layer to domain layer.
